Make things compile on macOS
