+++ Builds/Models/AppDelegate.swift
@@ -1,0 +1,74 @@
+// Copyright (c) 2022-2024 Jason Morley
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+import BackgroundTasks
+import SwiftUI
+
+#if os(iOS)
+
+class AppDelegate: NSObject, UIApplicationDelegate {
+
+    let applicationModel = ApplicationModel()
+
+    @MainActor func application(_ application: UIApplication,
+                                didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
+
+        // Background refresh handler.
+        BGTaskScheduler.shared.register(forTaskWithIdentifier: "uk.co.jbmorley.builds.tasks.refresh", using: nil) { task in
+            Task {
+                await self.applicationModel.refresh()
+                task.setTaskCompleted(success: true)
+            }
+        }
+
+        NotificationCenter.default.addObserver(forName: UIApplication.didEnterBackgroundNotification,
+                                               object: nil,
+                                               queue: nil) { _ in
+            DispatchQueue.main.async { [weak self] in
+                self?.scheduleBackgroundRefresh()
+            }
+        }
+
+        return true
+    }
+
+    func scheduleBackgroundRefresh() {
+        let request = BGAppRefreshTaskRequest(identifier: "uk.co.jbmorley.builds.tasks.refresh")
+        request.earliestBeginDate = Date(timeIntervalSinceNow: 15 * 60)
+        do {
+            print("Scheduling background refresh...")
+            try BGTaskScheduler.shared.submit(request)
+            print("Success!")
+        } catch {
+            print("Could not schedule app refresh: \(error)")
+        }
+    }
+
+}
+
+#else
+
+@MainActor class AppDelegate: NSObject, NSApplicationDelegate {
+
+    let applicationModel = ApplicationModel()
+
+}
+
+#endif
